Major update to map rendering and being able to finally explore.
Swapped the render method to the Camera class, and the populateView
method to the Viewport class. Added a WASD function to move around the
map inside the Camera class, rebuilding the viewport each time. May
adjust it to move up and down rather than diagonally. Stopped rendering
all tiles that were not visible on-screen. Did not seem to increase FPS.